% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidImage.R
\name{fluidImageOutput}
\alias{fluidImageOutput}
\title{Fluid image}
\usage{
fluidImageOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{id}

\item{width}{desired width}

\item{height}{desired height}
}
\description{
similar to \code{\link[shiny]{imageOutput}} but automtically scales to fit when resized.
}

