% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JcropImage.R
\name{JcropImageOutput}
\alias{JcropImageOutput}
\title{shiny image output widget with crop capabilities}
\usage{
JcropImageOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{name of the output}

\item{width}{desired width of the image output}

\item{height}{desired height of the image output}
}
\description{
shiny image output widget with crop capabilities powered by \code{Jcrop}.
As opposed to the usual \code{\link[shiny]{imageOutput}} component, this will
scale the image to fit the available space.
}
\details{
This emits these input events (for \code{outputId="image"}):
\itemize{
 \item{\code{image_change}}{
   When the selection changes. The event object contains coordinates.

This is also triggered when the image changes or when the cropper is released. In that case, the event
   object contains \code{NULL}
 }
 \item{\code{image_select}}{
   Same as \code{image_change}, but is triggered less often, only when the selection is fixed.
 }
}
}

