% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JcropImagePreview.R
\name{JcropImagePreview}
\alias{JcropImagePreview}
\title{Image preview widget}
\usage{
JcropImagePreview(image, event, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{image}{id of the associated \code{JcropImage} component}

\item{event}{input event. typically \code{image_change} or \code{image_select} if the
\code{JcropImage} output is called \code{image}}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{elementId}{unused}
}
\description{
Image preview widget that displays the selection from the \code{JcropImage} widget
}

