% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JcropImagePreview.R
\name{JcropImagePreviewOutput}
\alias{JcropImagePreviewOutput}
\title{Image preview output shiny component}
\usage{
JcropImagePreviewOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{name of the component}

\item{width}{width of component}

\item{height}{height of component}
}
\description{
Image preview output shiny component, suitable for a shiny ui.
}

