% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrpleCat.R
\name{purrple_logo}
\alias{purrple_logo}
\title{Adds the purrple logo at the bottom right}
\usage{
purrple_logo()
}
\description{
Adds the purrple logo at the bottom right of the application
The cat purrs when you mouse hover it
}

