% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrpleCat.R
\name{purrpleCat-shiny}
\alias{purrpleCat-shiny}
\alias{purrpleCatOutput}
\alias{renderPurrpleCat}
\title{Simple widget that adds the purrple logo at the right bottom corner}
\usage{
purrpleCatOutput(outputId)

renderPurrpleCat(session)
}
\arguments{
\item{outputId}{output variable to read from}

\item{session}{shiny session}
}
\description{
Simple widget that adds the purrple logo at the right bottom corner
}

