% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{purrplePage}
\alias{purrplePage}
\title{Tune a shiny page to add purrple branding (logo, ...)}
\usage{
purrplePage(fun)
}
\arguments{
\item{fun}{shiny page function (e.g. \code{\link[shiny]{fluidPage}}, ...)}
}
\value{
a modified version of the shiny page with additional html dependencies
}
\description{
Tune a shiny page to add purrple branding (logo, ...)
}

