% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidImage.R
\name{renderFluidImage}
\alias{renderFluidImage}
\title{renderer for fluid image output}
\usage{
renderFluidImage(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expression that evaluates to a file name on the server.}

\item{env}{environment}

\item{quoted}{whether the expression is quoted}
}
\description{
renderer for fluid image output
}

