% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JcropImage.R
\name{renderJcropImage}
\alias{renderJcropImage}
\title{Shiny renderer for JcropImageOutput}
\usage{
renderJcropImage(expr, background = "white", opacity = 0.3,
  aspect_ratio = 1, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expression that evaluates to a file name on the server. typically a temporary file.}

\item{background}{background color to use when the selection is active}

\item{opacity}{opacity of the background}

\item{aspect_ratio}{aspect ratio (width / height) of the selection. If \code{NULL} the selection is free.}

\item{env}{environment in which to evaluate \code{expr}}

\item{quoted}{is the expression quoted}
}
\description{
Shiny renderer for JcropImageOutput
}

