% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JcropImagePreview.R
\name{renderJcropImagePreview}
\alias{renderJcropImagePreview}
\title{Renderer for Image Preview component}
\usage{
renderJcropImagePreview(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expression to evaluate, should end with a call to \code{\link{JcropImagePreview}}}

\item{env}{environment where to evaluate \code{expr}}

\item{quoted}{Is the expression already quoted}
}
\description{
Renderer for Image Preview component
}

