% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{basicPage}
\alias{basicPage}
\alias{bootstrapPage}
\alias{fillPage}
\alias{fixedPage}
\alias{fluidPage}
\alias{navbarPage}
\title{Modified version of shiny pages with purrple branding}
\usage{
basicPage(...)

bootstrapPage(...)

fillPage(...)

fixedPage(...)

navbarPage(...)

fluidPage(...)
}
\arguments{
\item{\dots}{see \link[shiny]{basicPage}}
}
\description{
Modified version of shiny pages with purrple branding
}

