% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidImage.R
\name{renderFluidImage}
\alias{renderFluidImage}
\title{renderer for fluid image output}
\usage{
renderFluidImage(expr, session = get("session", parent.frame()))
}
\arguments{
\item{expr}{expression that evaluates to a file name on the server.}

\item{session}{shiny session}
}
\description{
renderer for fluid image output
}

